:art: FLOW-647 helpers updated
